added explicit casting

+++ src/twopi/vkl/primitive/vkl_floor.h
@@ -16,8 +16,8 @@ public:
 
   ~Floor();
 
-  const int NumVertices() const { return position_buffer_.size() / 3; }
-  const int NumIndices() const { return index_buffer_.size(); }
+  const int NumVertices() const { return static_cast<int>(position_buffer_.size() / 3); }
+  const int NumIndices() const { return static_cast<int>(index_buffer_.size()); }
 
   const auto& PositionBuffer() const { return position_buffer_; }
   const auto& NormalBuffer() const { return normal_buffer_; }
